backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tekno-phantom-backend"
version = "0.1.0"
description = "FastAPI backend for Tekno Phantom Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "mcp>=1.26.0",
  "playwright>=1.49.0",
  "pillow>=10.4.0",
  "python-multipart>=0.0.9",
  "openpyxl>=3.1.0",
  "sqlalchemy>=2.0.35",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt>=4.0.1,<4.1",
  "python-jose[cryptography]>=3.3.0",
  "email-validator>=2.2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["artifacts*", "tests*"]
